requirements.txt
# Core dependencies
python-dotenv>=1.0.0
pyyaml>=6.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# LLM providers
openai>=1.10.0
anthropic>=0.18.0

# Exchange APIs
ccxt>=4.2.0
python-binance>=1.0.19

# Data processing
numpy>=1.26.0
pandas>=2.1.0
ta>=0.11.0  # Technical analysis library

# Async support
aiohttp>=3.9.0
asyncio>=3.4.3

# Database (for logging and state)
sqlalchemy>=2.0.0
alembic>=1.13.0

# Monitoring and logging
prometheus-client>=0.19.0
structlog>=24.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Utilities
requests>=2.31.0
python-dateutil>=2.8.2
pytz>=2023.3
tenacity>=8.2.0  # Retry logic

# Web interface (optional)
fastapi>=0.109.0
uvicorn>=0.27.0
